+++ ft_utilities_hex_uns.c
@@ -6,7 +6,7 @@
 /*   By: lpaiva <lpaiva@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/07 23:42:28 by lpaiva            #+#    #+#             */
-/*   Updated: 2025/11/10 18:40:33 by lpaiva           ###   ########.fr       */
+/*   Updated: 2025/11/12 22:33:28 by lpaiva           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,10 +16,10 @@
 //  for unsignd int and hexadecimal numbers and pointrs
 #include "ft_printf.h"
 
-int	ft_putnbr_base_uns(unsigned long long nbr, char *base, int base_len,
+int	ft_putnbr_base_uns(unsigned long nbr, char *base, int base_len,
 		int count)
 {
-	if (nbr >= (unsigned long long)base_len)
+	if (nbr >= (unsigned long)base_len)
 		count = ft_putnbr_base_uns(nbr / base_len, base, base_len, count);
 	count = ft_putchar(base[nbr % base_len], count);
 	return (count);
